Add minor-mode and remove extra progn forms melpa/melpa#7570
As recommended in the MELPA pull-request, I have reworked this package
to only modify the behavior of langtool when the
langtool-ignore-fonts-minor-mode is enabled.

I have removed the extra progn forms.

I have also added the setter function langtool-ignore-fonts-add for
setting which fonts langtool should ignore in a given major mode.